Save active directory or file in properties file
I will create a $USER/.sgf4j-gui folder with a properties file inside.
sgf4j-gui.properties. I'm storing the active directory there so that on
subsequen starts of the app the previous folder is navigated back to.

Right now we are still missing the restoring of the saved state in app.
Should be an easy add.