Some programming by coincidence
Not sure why the circle and highlight circle look so off when resizing and sometimes at random. My debugging with ScenicView shows that the LayoutX and LayoutY are quite different for them. So I'm trying to set them to be the same.

It has improved the visual on my Windows machine. I'll try to verify the changes on Linux and Mac also. It doesn't seem to fix everything but right now on Windows it fixes a lot.